Rename networkattrs to networkattributes
