+++ core/src/main/java/io/opentelemetry/android/internal/features/networkattrs/NetworkAttributesSpanAppender.java
@@ -5,8 +5,8 @@
 
 package io.opentelemetry.android.internal.features.networkattrs;
 
-import io.opentelemetry.android.common.internal.features.networkattrs.CurrentNetworkAttributesExtractor;
-import io.opentelemetry.android.common.internal.features.networkattrs.data.CurrentNetwork;
+import io.opentelemetry.android.common.internal.features.networkattributes.CurrentNetworkAttributesExtractor;
+import io.opentelemetry.android.common.internal.features.networkattributes.data.CurrentNetwork;
 import io.opentelemetry.android.internal.services.network.CurrentNetworkProvider;
 import io.opentelemetry.api.common.Attributes;
 import io.opentelemetry.context.Context;
