+++ services/src/main/java/io/opentelemetry/android/internal/services/network/CurrentNetworkProvider.java
@@ -15,8 +15,8 @@ import android.os.Build;
 import android.util.Log;
 import androidx.annotation.NonNull;
 import io.opentelemetry.android.common.RumConstants;
-import io.opentelemetry.android.common.internal.features.networkattrs.data.CurrentNetwork;
-import io.opentelemetry.android.common.internal.features.networkattrs.data.NetworkState;
+import io.opentelemetry.android.common.internal.features.networkattributes.data.CurrentNetwork;
+import io.opentelemetry.android.common.internal.features.networkattributes.data.NetworkState;
 import io.opentelemetry.android.internal.services.Startable;
 import io.opentelemetry.android.internal.services.network.detector.NetworkDetector;
 import java.util.List;
