+++ services/src/main/java/io/opentelemetry/android/internal/services/network/detector/NetworkDetector.java
@@ -9,7 +9,7 @@ import android.content.Context;
 import android.net.ConnectivityManager;
 import android.os.Build;
 import android.telephony.TelephonyManager;
-import io.opentelemetry.android.common.internal.features.networkattrs.data.CurrentNetwork;
+import io.opentelemetry.android.common.internal.features.networkattributes.data.CurrentNetwork;
 import io.opentelemetry.android.internal.services.network.CarrierFinder;
 
 /**
