+++ services/src/test/java/io/opentelemetry/android/internal/services/network/CurrentNetworkProviderTest.java
@@ -21,8 +21,8 @@ import android.net.ConnectivityManager.NetworkCallback;
 import android.net.Network;
 import android.net.NetworkRequest;
 import android.os.Build;
-import io.opentelemetry.android.common.internal.features.networkattrs.data.CurrentNetwork;
-import io.opentelemetry.android.common.internal.features.networkattrs.data.NetworkState;
+import io.opentelemetry.android.common.internal.features.networkattributes.data.CurrentNetwork;
+import io.opentelemetry.android.common.internal.features.networkattributes.data.NetworkState;
 import io.opentelemetry.android.internal.services.network.detector.NetworkDetector;
 import java.util.concurrent.atomic.AtomicInteger;
 import org.junit.Assert;
