+++ services/src/test/java/io/opentelemetry/android/internal/services/network/detector/SimpleNetworkDetectorTest.java
@@ -13,8 +13,8 @@ import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
 import android.os.Build;
 import androidx.test.core.app.ApplicationProvider;
-import io.opentelemetry.android.common.internal.features.networkattrs.data.CurrentNetwork;
-import io.opentelemetry.android.common.internal.features.networkattrs.data.NetworkState;
+import io.opentelemetry.android.common.internal.features.networkattributes.data.CurrentNetwork;
+import io.opentelemetry.android.common.internal.features.networkattributes.data.NetworkState;
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
